Adds the ability to prevent source code folders from being deleted
Prior to this PR, the buildpack would delete all source code (i.e. everything under `/workspace`). This is done because Rust compiles to a binary and so you don't need the source code to actually run the application.

This overlooked some cases where an application might need or want to include files in the image other than the binary. The classic example would be static Javascript/CSS/HTML, or template files used by a web application.

This PR adds a configuration option called `$BP_CARGO_EXCLUDE_FOLDERS` which is a comma separate list of folders that should not be deleted. This list must be top-level folders only (i.e. what is directly under `/workspace`). It does not work with nested folders. The default is `static, templates, public, html` which attempts to cover some of the standard and common locations for static and template files.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>