Add `tini` to launch processes
By default, the buildpack will install `tini` and configure process types to use `tini` to start your application. Your process runs as PID1, so this ensures that signals will work properly out-of-the-box. The `tini` binary is very small so this does not add a lot of size or overhead. If you do not need this behavior, you can disable it by setting `$BP_CARGO_TINI_DISABLED` to `true`.

This is installing `tini` directly instead of using the `tini` buildpack for a couple of reasons. First, we needed to make changes to this buildpack to support `tini` anyway, it is a only small bit of code, it ensures that output is consistent, and when we get to SBOM it will ensure that operates consistently with the rest of the buildpack.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>