Automatic build on the Tiny stack
Prior to this PR, you would need to manually adjust cargo build argumets to build properly on the Tiny stack. This is a requirement because to run on the Tiny stack you need to compile against musl not glibc.

If running on the tiny stack, this PR will automatically add the cargo build arguments `--target x86_64-unknown-linux-musl`. If the `--target` has already been set, it does not change the user provided value. If not running on the tiny stack, no `--target` is added.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>