Store SBOM scan output for application on the layer
SBOM that is stored on the top-level build/launch location is not restored, so it must be regenerated every time. That is a problem because syft can take multiple seconds to scan some larger Rust projects. As such, we are storing the output on the layer instead. The SBOM information is restored when it is associated with a layer. We are storing it on the Cargo layer, which is named "Rust Application" in the build output, so that logically makes sense too.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>