Partial Support for RFC #44 - Disable SBOM Generation
This commit adds partial support for disabling SBOM generation by setting `BP_DISABLE_SBOM=true`. This will presently disable the expensive operation of scanning your application. It does not disable the cheap operation of adding SBOM for dependencies that are installed by the buildpack. Support to disable that will require a change to libpak.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>