Allow installing rust tools
This provides an env variable `$BP_CARGO_INSTALL_TOOLS` which can be given a space separated list of tools that should be installed using `cargo install`. In addition, it provides `$BP_CARGO_INSTALL_TOOLS_ARGS` which can be used to supply arguments to `cargo install` when used to install tools.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>