Switch to using RFC Utilities RFC #0005
Utilities RFC 0005 defines how source file deletion should be performed, what key environment variables should be considered, provides the core logic and a buildpack which implements the RFC.

This PR utilizes the logic contributed by RFC 0005 as a Go package and the environment variables defined in it to specify included/excluded files.

It also deprecates using the 'BP_CARGO_EXCLUDE_FOLDERS' environment variable. If this environment variable is used, the values in it are converted to the format expected by the RFC 0005 implementation, so it should not be a breaking change. A warning message will also be emitted. 'BP_CARGO_EXCLUDE_FOLDERS' will be removed before the 1.0 GA release.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>