Enable sparse registry support.
Cargo has [recently enabled what is called sparse registry support](https://blog.rust-lang.org/2022/06/22/sparse-registry-testing.html) on nightly Rust. This will download crates.io information over HTTPS instead of Git. It is much faster because it's also able to only download what it needs for the project instead of the entire set of data from crates.io. See the link for more details.

This PR sets the environment variable to enable sparse registry support, which can decrease buildtimes significantly on the first run. This is only supported on nightly Rust at the moment, but the environment variable is safely ignored on stable. When it's supported on stable, it should just start working there as well.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>