Change target selection to support ARM64
Previously, the target selection on the tiny/static stacks would assume AMD64/x86_64. This PR changes the behavior to look at the current system architecture and select the correct target for the architecture running. It also enables the `BP_ARCH` environment variable which can be used to override the decision, although this is primarily for testing.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>